+++ crypto/secp256k1/secp256k1_nocgo_test.go
@@ -1,4 +1,4 @@
-// +build !libsecp256k1 unit
+// +build !libsecp256k1,unit
 
 package secp256k1
